requirements.txt
aiogram>=3.0.0
Flask>=2.2.0
flask-socketio>=5.3.6
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=1.5.3
matplotlib>=3.7.0
mplfinance==0.12.10b0
numpy>=1.23.5
setuptools>=65.5.0
aiohttp>=3.9.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0